Separate Turrets file for North Africa
