+++ addons/african_north/CfgVehicles_Turrets.hpp
@@ -1,0 +1,94 @@
+// Turrets - Blufor
+class TACU_African_North_V_B_M2HMG_Low: O_SFIA_HMG_02_lxWS {
+    faction = "TACU_African_North_B";
+    side = 1;
+    scope = 2;
+    scopeCurator = 2;
+    crew = "TACU_African_North_U_B_Repair";
+    typicalCargo[] = {"TACU_African_North_U_B_Repair"};
+};
+
+class TACU_African_North_V_B_M2HMG: O_SFIA_HMG_02_high_lxWS {
+    faction = "TACU_African_North_B";
+    side = 1;
+    scope = 2;
+    scopeCurator = 2;
+    crew = "TACU_African_North_U_B_Repair";
+    typicalCargo[] = {"TACU_African_North_U_B_Repair"};
+};
+
+class TACU_African_North_V_B_MK6: O_SFIA_Mortar_lxWS {
+    faction = "TACU_African_North_B";
+    side = 1;
+    scope = 2;
+    scopeCurator = 2;
+    crew = "TACU_African_North_U_B_Repair";
+    typicalCargo[] = {"TACU_African_North_U_B_Repair"};
+};
+
+class TACU_African_North_V_B_ZU23: O_SFIA_ZU23_lxWS {
+    faction = "TACU_African_North_B";
+    side = 1;
+    scope = 2;
+    scopeCurator = 2;
+    crew = "TACU_African_North_U_B_Repair";
+    typicalCargo[] = {"TACU_African_North_U_B_Repair"};
+};
+
+// Turrets - Opfor
+class TACU_African_North_V_O_M2HMG_Low: TACU_African_North_V_B_M2HMG_Low {
+    faction = "TACU_African_North_O";
+    side = 0;
+    crew = "TACU_African_North_U_O_Repair";
+    typicalCargo[] = {"TACU_African_North_U_O_Repair"};
+};
+
+class TACU_African_North_V_O_M2HMG: TACU_African_North_V_B_M2HMG {
+    faction = "TACU_African_North_O";
+    side = 0;
+    crew = "TACU_African_North_U_O_Repair";
+    typicalCargo[] = {"TACU_African_North_U_O_Repair"};
+};
+
+class TACU_African_North_V_O_MK6: TACU_African_North_V_B_MK6 {
+    faction = "TACU_African_North_O";
+    side = 0;
+    crew = "TACU_African_North_U_O_Repair";
+    typicalCargo[] = {"TACU_African_North_U_O_Repair"};
+};
+
+class TACU_African_North_V_O_ZU23: TACU_African_North_V_B_ZU23 {
+    faction = "TACU_African_North_O";
+    side = 0;
+    crew = "TACU_African_North_U_O_Repair";
+    typicalCargo[] = {"TACU_African_North_U_O_Repair"};
+};
+
+// Turrets - Independent
+class TACU_African_North_V_I_M2HMG_Low: TACU_African_North_V_B_M2HMG_Low {
+    faction = "TACU_African_North_I";
+    side = 2;
+    crew = "TACU_African_North_U_I_Repair";
+    typicalCargo[] = {"TACU_African_North_U_I_Repair"};
+};
+
+class TACU_African_North_V_I_M2HMG: TACU_African_North_V_B_M2HMG {
+    faction = "TACU_African_North_I";
+    side = 2;
+    crew = "TACU_African_North_U_I_Repair";
+    typicalCargo[] = {"TACU_African_North_U_I_Repair"};
+};
+
+class TACU_African_North_V_I_MK6: TACU_African_North_V_B_MK6 {
+    faction = "TACU_African_North_I";
+    side = 2;
+    crew = "TACU_African_North_U_I_Repair";
+    typicalCargo[] = {"TACU_African_North_U_I_Repair"};
+};
+
+class TACU_African_North_V_I_ZU23: TACU_African_North_V_B_ZU23 {
+    faction = "TACU_African_North_I";
+    side = 2;
+    crew = "TACU_African_North_U_I_Repair";
+    typicalCargo[] = {"TACU_African_North_U_I_Repair"};
+};
