+++ addons/african_north/CfgVehicles_Vehicles_B.hpp
@@ -235,40 +235,3 @@ class TACU_African_North_V_B_Offroad_HMG: O_SFIA_Offroad_armed_lxWS {
     crew = "TACU_African_North_U_B_Repair";
     typicalCargo[] = {"TACU_African_North_U_B_Repair"};
 };
-
-// Turrets
-class TACU_African_North_V_B_M2HMG_Low: O_SFIA_HMG_02_lxWS {
-    faction = "TACU_African_North_B";
-    side = 1;
-    scope = 2;
-    scopeCurator = 2;
-    crew = "TACU_African_North_U_B_Repair";
-    typicalCargo[] = {"TACU_African_North_U_B_Repair"};
-};
-
-class TACU_African_North_V_B_M2HMG: O_SFIA_HMG_02_high_lxWS {
-    faction = "TACU_African_North_B";
-    side = 1;
-    scope = 2;
-    scopeCurator = 2;
-    crew = "TACU_African_North_U_B_Repair";
-    typicalCargo[] = {"TACU_African_North_U_B_Repair"};
-};
-
-class TACU_African_North_V_B_MK6: O_SFIA_Mortar_lxWS {
-    faction = "TACU_African_North_B";
-    side = 1;
-    scope = 2;
-    scopeCurator = 2;
-    crew = "TACU_African_North_U_B_Repair";
-    typicalCargo[] = {"TACU_African_North_U_B_Repair"};
-};
-
-class TACU_African_North_V_B_ZU23: O_SFIA_ZU23_lxWS {
-    faction = "TACU_African_North_B";
-    side = 1;
-    scope = 2;
-    scopeCurator = 2;
-    crew = "TACU_African_North_U_B_Repair";
-    typicalCargo[] = {"TACU_African_North_U_B_Repair"};
-};
